# Conflicts:
#	.gitignore
#	PluginSource/Assets/Editor/IOSBuildPostProcess.cs
#	PluginSource/Assets/Plugins/iOS/Spil.framework/FYBBackIcon@2x.png.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/FYBBackIconiPad@2x.png.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/FYBBackIconiPhone@2x.png.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Frameworks/Chartboost.framework/Chartboost.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Frameworks/FBAudienceNetwork.framework/FBAudienceNetwork.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Frameworks/FBAudienceNetwork.framework/Info.plist.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Frameworks/Fyber_AdColony.framework/Fyber_AdColony.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Frameworks/Fyber_AdColony.framework/Resources/Info.plist.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Frameworks/Fyber_AppLovin.framework/Fyber_AppLovin.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Frameworks/Fyber_AppLovin.framework/Resources/Info.plist.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Frameworks/Fyber_UnityAds.framework/Fyber_UnityAds.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Frameworks/Fyber_UnityAds.framework/Resources/Info.plist.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Frameworks/Fyber_UnityAds.framework/Resources/UnityAds.bundle/Info.plist.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Frameworks/Fyber_UnityAds.framework/Resources/UnityAds.bundle/audio_mute.png.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Frameworks/Fyber_UnityAds.framework/Resources/UnityAds.bundle/audio_mute@2x.png.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Frameworks/Fyber_UnityAds.framework/Resources/UnityAds.bundle/audio_on.png.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Frameworks/Fyber_UnityAds.framework/Resources/UnityAds.bundle/audio_on@2x.png.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Frameworks/Fyber_UnityAds.framework/Resources/UnityAds.bundle/en.lproj/InfoPlist.strings.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Frameworks/Fyber_UnityAds.framework/Resources/UnityAdsCompiler-Info.plist.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Frameworks/GoogleMobileAds.framework/GoogleMobileAds.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Frameworks/MMAdSDK.framework/MMAdSDK.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Frameworks/libAdMobMMAdapter.a.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Frameworks/libAdapterFacebook.a.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Frameworks/libAdapterSDKMoPub.a.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Frameworks/libFyberSDK.a.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Headers/HookBridge.h.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Headers/JsonUtil.h.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Headers/Spil.h.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Headers/UnityAppController.h.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Info.plist.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/LICENSE.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/MPAdBrowserController.nib
#	PluginSource/Assets/Plugins/iOS/Spil.framework/MPAdBrowserController.nib.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/MPCloseBtn.png.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/MPCloseBtn@2x.png.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/MPCloseBtn@3x.png.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/MPCloseButtonX.png.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/MPCloseButtonX@2x.png.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/MPCloseButtonX@3x.png.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/MPDAAIcon.png.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/MPDAAIcon@2x.png.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/MPDAAIcon@3x.png.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/MPMutedBtn.png.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/MPMutedBtn@2x.png.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/MPMutedBtn@3x.png.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/MPPlayBtn.png.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/MPPlayBtn@2x.png.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/MPPlayBtn@3x.png.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/MPUnmutedBtn.png.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/MPUnmutedBtn@2x.png.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/MPUnmutedBtn@3x.png.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/MRAID.bundle/mraid.jjs.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Modules/module.modulemap.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Settings.bundle/DebugMode.plist.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Settings.bundle/Root.plist.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Settings.bundle/en.lproj/Root.strings.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Spil
#	PluginSource/Assets/Plugins/iOS/Spil.framework/Spil.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/_CodeSignature/CodeDirectory
#	PluginSource/Assets/Plugins/iOS/Spil.framework/_CodeSignature/CodeDirectory.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/_CodeSignature/CodeRequirements-1
#	PluginSource/Assets/Plugins/iOS/Spil.framework/_CodeSignature/CodeRequirements-1.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/_CodeSignature/CodeRequirements.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/_CodeSignature/CodeResources
#	PluginSource/Assets/Plugins/iOS/Spil.framework/_CodeSignature/CodeResources.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/_CodeSignature/CodeSignature
#	PluginSource/Assets/Plugins/iOS/Spil.framework/_CodeSignature/CodeSignature.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/_CodeSignature/CodeTopDirectory.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/defaultGameData.json.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/defaultPlayerData.json.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/mod_pbxproj.py.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/project.entitlements.meta
#	PluginSource/Assets/Plugins/iOS/Spil.framework/setup.py.meta

+++ PluginSource/Assets/Plugins/iOS/Spil.framework/setup.py
@@ -18,12 +18,19 @@ BOLD = '\033[1m'
 
 # helper methods
 def addBundleResource(src, dst, group):
+	# check if the resource actually exists
+	if not os.path.exists(src):
+		return
+
+	# copy the dir or file
 	if os.path.isdir(src):
 		if os.path.exists(dst):
 			shutil.rmtree(dst)
 		shutil.copytree(src, dst)
 	else:
-		shutil.copy2(src, dst)	
+		shutil.copy2(src, dst)
+	
+	# add it to the xcode project	
 	project.add_file_if_doesnt_exist(dst, parent=bundles, weak=False)
 	return
 
@@ -101,6 +108,9 @@ bundles = project.get_or_create_group('')
 addBundleResource(os.getcwd() + '/Spil.framework/Settings.bundle', os.getcwd() + '/Settings.bundle', bundles)
 addBundleResource(os.getcwd() + '/Spil.framework/Frameworks/Fyber_UnityAds.framework/Resources/UnityAds.bundle', os.getcwd() + '/UnityAds.bundle', bundles)
 addBundleResource(os.getcwd() + '/Spil.framework/project.entitlements', os.getcwd() + '/' + projectname + '.entitlements', bundles)
+addBundleResource(os.getcwd() + '/Data/Raw/defaultGameConfig.json', os.getcwd() + '/defaultGameConfig.json', bundles)
+addBundleResource(os.getcwd() + '/Data/Raw/defaultGameData.json', os.getcwd() + '/defaultGameData.json', bundles)
+addBundleResource(os.getcwd() + '/Data/Raw/defaultPlayerData.json', os.getcwd() + '/defaultPlayerData.json', bundles)
 
 # change build settings
 print 'Modifying project build settings'
